% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parasiteModel.r
\name{parasiteModel}
\alias{parasiteModel}
\title{Calculates regression from count data}
\usage{
parasiteModel(farebna.paleta = "Accent")
}
\arguments{
\item{farebna.paleta}{character. Color palette in \code{hcl.pals()} or
\code{palette.pals()}.}
}
\value{
Called for side effects. Stores summary of the model fit in a text file and
plots predicted values in pdf files.
}
\description{
Fits generalized linear model with negative binomial distribution suitable for
analysing count data such as number of ectoparasites. Allows the user to choose
working directory and the data file, choose the dependent and explanatory variables,
fits the model and plots variable predictions as specified by the user.
}
\details{
Function communicates with the user in Slovak to distinguish what dialogs
represent intended functionality. Pay attention to dialog boxes (2) at the beginning
and prompts on the R console afterwards.
}
