% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleModel.r
\name{simpleModel}
\alias{simpleModel}
\title{Calculates linear regression model from continuous data}
\usage{
simpleModel(farebna.paleta = "Accent", vymaz.odlahle = FALSE)
}
\arguments{
\item{farebna.paleta}{character. Color palette in \code{hcl.pals()} or
\code{palette.pals()}.}

\item{vymaz.odlahle}{logical, indicating whether to remove outliers in the dependent
variable using Tukey's fences.}
}
\value{
Called for side effects. Stores summary of the model fit in a text file and
plots predicted values in pdf files.
}
\description{
Fits linear model for
analysing continuous data, where one measurement was taken per animal on
each day, such as weight or body temperature. Allows the user to choose
working directory and the data file, choose the dependent and explanatory variables,
fits the model and plots variable predictions as specified by the user.
}
\details{
Function communicates with the user in Slovak to distinguish what dialogs
represent intended functionality. Pay attention to dialog boxes (2) at the beginning
and prompts on the R console afterwards.

Dialogs:
\enumerate{
\item Select a folder where the results will be saved.
\item Select a tab-delimited table with data describing the animals. This was likely
provided by the supervisor.
}

Once the data tables are selected, the function asks to select variables that will
be used to construct the model. The predicted values from the model can be plotted,
displaying one partial fit at the time. All results can be found in the folder
selected in dialog 1.
}
\examples{
\dontrun{
# Open R console and type
simpleModel()
# Hit Enter and follow instructions in dialogs and on the console
}
}
