% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeFromSunrise.r
\name{timeFromSunrise}
\alias{timeFromSunrise}
\title{Time from Sunrise}
\usage{
timeFromSunrise(hour, minute, date, site)
}
\arguments{
\item{hour}{numeric. Hour of capture.}

\item{minute}{numeric. Minute of capture.}

\item{date}{Date. Date of capture.}

\item{site}{character. Site name with latitude and longitude in decimal degrees.}
}
\value{
Number of hours since sunrise when the animal was captured.
}
\description{
Calculates how many hours after sunrise the animal was caught. Sunrise at the site
is calculated from the geographic coordinates of the site description.
}
\details{
At this time, only geographic coordinates on the northern and eastern
hemispheres are considered.
}
\examples{
  timeFromSunrise(hour = 10, minute = 30, date = as.Date("2023-05-30"), site = "Hustopece, CZ, 48.93N, 16.72E")
}
